Added GroupClient
