+++ pyatlan/client/audit.py
@@ -1,7 +1,7 @@
 from pydantic import ValidationError, parse_obj_as
 
+from pyatlan.client.common import ApiCaller
 from pyatlan.client.constants import AUDIT_SEARCH
-from pyatlan.client.workflow import ApiCaller
 from pyatlan.errors import ErrorCode
 from pyatlan.model.audit import AuditSearchRequest, AuditSearchResults, EntityAudit
